+++ docs/source/changelog.rst
@@ -7,6 +7,14 @@ This project adheres to `Semantic Versioning`_.
 
 .. _`Semantic Versioning`: https://semver.org/spec/v2.0.0.html
 
+
+0.3.0
+=====
+
+  - Added keyword only flag for Value and Configuration classes.
+  - Added cache option for Configuration class.
+
+
 0.2.0
 =====